refactor: comment 패키지 studyroom으로 이동
