--- backend/src/main/java/com/woowacourse/moamoa/comment/domain/AssociatedCommunity.java
@@ -1,21 +1,0 @@
-package com.woowacourse.moamoa.comment.domain;
-
-import static lombok.AccessLevel.PROTECTED;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@NoArgsConstructor(access = PROTECTED)
-@AllArgsConstructor
-@Getter
-@EqualsAndHashCode
-public class AssociatedCommunity {
-
-    @Column(name = "community_id", nullable = false)
-    private Long communityId;
-}
