--- backend/src/main/java/com/woowacourse/moamoa/comment/domain/Author.java
@@ -1,20 +1,0 @@
-package com.woowacourse.moamoa.comment.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EqualsAndHashCode
-@AllArgsConstructor
-public class Author {
-
-    @Column(name = "author_id", nullable = false)
-    private Long authorId;
-}
