--- backend/src/main/java/com/woowacourse/moamoa/comment/domain/repository/JpaCommentRepository.java
@@ -1,7 +1,0 @@
-package com.woowacourse.moamoa.comment.domain.repository;
-
-import com.woowacourse.moamoa.comment.domain.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface JpaCommentRepository  extends JpaRepository<Comment, Long>, CommentRepository {
-}
