--- backend/src/main/java/com/woowacourse/moamoa/comment/service/exception/UnEditingCommentException.java
@@ -1,10 +1,0 @@
-package com.woowacourse.moamoa.comment.service.exception;
-
-import com.woowacourse.moamoa.common.exception.BadRequestException;
-
-public class UnEditingCommentException extends BadRequestException {
-
-    public UnEditingCommentException() {
-        super("댓글을 삭제할 수 없습니다.");
-    }
-}
