--- backend/src/main/java/com/woowacourse/moamoa/comment/service/exception/UnWrittenCommentException.java
@@ -1,10 +1,0 @@
-package com.woowacourse.moamoa.comment.service.exception;
-
-import com.woowacourse.moamoa.common.exception.BadRequestException;
-
-public class UnWrittenCommentException extends BadRequestException {
-
-    public UnWrittenCommentException() {
-        super("댓글을 작성하거나 수정할 수 없습니다.");
-    }
-}
