+++ backend/src/main/java/com/woowacourse/moamoa/studyroom/domain/comment/AssociatedCommunity.java
@@ -1,0 +1,21 @@
+package com.woowacourse.moamoa.studyroom.domain.comment;
+
+import static lombok.AccessLevel.PROTECTED;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@NoArgsConstructor(access = PROTECTED)
+@AllArgsConstructor
+@Getter
+@EqualsAndHashCode
+public class AssociatedCommunity {
+
+    @Column(name = "community_id", nullable = false)
+    private Long communityId;
+}
