+++ backend/src/main/java/com/woowacourse/moamoa/studyroom/domain/comment/repository/JpaCommentRepository.java
@@ -1,0 +1,7 @@
+package com.woowacourse.moamoa.studyroom.domain.comment.repository;
+
+import com.woowacourse.moamoa.studyroom.domain.comment.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface JpaCommentRepository  extends JpaRepository<Comment, Long>, CommentRepository {
+}
