+++ backend/src/main/java/com/woowacourse/moamoa/studyroom/service/request/CommentRequest.java
@@ -1,0 +1,17 @@
+package com.woowacourse.moamoa.studyroom.service.request;
+
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.validator.constraints.Length;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CommentRequest {
+
+    @NotBlank(message = "내용을 입력해 주세요.")
+    @Length(max = 5000)
+    private String content;
+}
