+++ backend/src/main/java/com/woowacourse/moamoa/studyroom/service/request/EditingCommentRequest.java
@@ -1,0 +1,15 @@
+package com.woowacourse.moamoa.studyroom.service.request;
+
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class EditingCommentRequest {
+
+    @NotBlank(message = "내용을 입력해 주세요.")
+    private String content;
+}
